feat: use GSON to access movies data
